Django_Day11_workshop

+++ Web/SS4th/7th_Django/CRUD/articles/views.py
@@ -22,7 +22,9 @@ def create(request):
     if request.method == 'POST':
         form = ArticleForm(request.POST, request.FILES)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
         # article = Article()
         # article.title = request.POST.get('title')
         # article.content = request.POST.get('content')
@@ -59,29 +61,34 @@ def detail(request, pk):
 @login_required
 def update(request, pk):
     article = Article.objects.get(pk=pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, request.FILES, instance=article)
-        if form.is_valid():
-            article = form.save()
-            # article.title = request.POST.get('title')
-            # article.content = request.POST.get('content')
-            # article.save()
-            return redirect('articles:detail', article.pk)
+    if request.user == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, request.FILES, instance=article)
+            if form.is_valid():
+                article = form.save()
+                # article.title = request.POST.get('title')
+                # article.content = request.POST.get('content')
+                # article.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
+        return redirect('articles:detail', article.pk)
     context = {
         'article': article,
         'form': form
     }
     return render(request, 'articles/update.html', context)
 
-
-@require_POST
+@login_required
+# @require_POST
 def delete(request, pk):
     article = Article.objects.get(pk=pk)
-    # if request.method == 'POST':
-    article.delete()
-    return redirect('articles:index')
+    if request.method == 'POST':
+        if request.user == article.user:
+            article.delete()
+            return redirect('articles:index')
+    return redirect('articles:detail', article.pk)
     # else:
     #     return redirect('articles:detail', article.pk)
 
@@ -93,17 +100,21 @@ def comment_create(request, pk):
     if comment_form.is_valid():
         comment = comment_form.save(commit=False)
         comment.article = article
+        comment.user = request.user
         comment.save()
         return redirect('articles:detail', article.pk)
     context = {
         'comment_form': comment_form,
         'article': article,
     }
-    return render(request, 'articles:detail', article.pk)
+    return render(request, 'articles/detail.html', context)
 
 
-@require_POST
+@login_required
+# @require_POST
 def comment_delete(request, article_pk, comment_pk):
     comment = get_object_or_404(Comment, pk=comment_pk)
-    comment.delete()
+    if request.user == comment.user:
+        if request.method == 'POST':
+            comment.delete()
     return redirect('articles:detail', article_pk)
